requirements.txt
fastapi
uvicorn
openai
tavily-python
